Revert "decals just in case"
This reverts commit 1088ddc202a80f51680b4f339eb200b23433f1b7.

+++ Assets/!Materials/Projectors/ImageProjector.shader
@@ -19,19 +19,19 @@
  
          // User-specified properties
          uniform sampler2D _ShadowTex; 
-         fixed4 _Color;
+         float4 _Color;
  
          // Projector-specific uniforms
-         uniform fixed4x4 _Projector; // transformation matrix 
+         uniform float4x4 _Projector; // transformation matrix 
             // from object space to projector space 
  
           struct vertexInput {
-            fixed4 vertex : POSITION;
+            float4 vertex : POSITION;
             float3 normal : NORMAL;
          };
          struct vertexOutput {
-            fixed4 pos : SV_POSITION;
-            fixed4 posProj : TEXCOORD0;
+            float4 pos : SV_POSITION;
+            float4 posProj : TEXCOORD0;
                // position in projector space
          };
  
@@ -45,12 +45,12 @@
          }
  
  
-         fixed4 frag(vertexOutput input) : COLOR
+         float4 frag(vertexOutput input) : COLOR
          {
             if (input.posProj.w > 0.0) // in front of projector?
             {
                //return tex2D(_ShadowTex , input.posProj.xy / input.posProj.w); 
-               fixed4 f = tex2D(_ShadowTex , input.posProj.xy / input.posProj.w);
+               float4 f = tex2D(_ShadowTex , input.posProj.xy / input.posProj.w);
                f.rgb *= _Color.rgb;
                return f;
 
@@ -59,7 +59,7 @@
             }
             else // behind projector
             {
-               return fixed4(0.0, 0.0, 0.0, 0.0);
+               return float4(0.0, 0.0, 0.0, 0.0);
             }
          }
